requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
antiorm==1.2.1
attrs==25.1.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
databases==0.9.0
db==0.1.1
discord==2.3.2
discord.py==2.5.2
dominate==2.9.1
expiringdict==1.2.2
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
flask-cors==5.0.1
Flask-Discord==0.1.69
Flask-WTF==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markup==0.2.2
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
visitor==0.1.3
waitress==3.0.2
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
